# backend/openapi.yaml
openapi: 3.0.3
info:
  title: SPV API
  version: 1.0.0
  description: API para votos, puntos y analisis de planificacion.
servers:
  - url: http://localhost:8787
paths:
...
  /api/identity/link:
    post:
      summary: Vincular usuario externo de una app con usuario interno SVP
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [sourceApp, externalUserId]
              properties:
                sourceApp:
                  type: string
                externalUserId:
                  type: string
                svpUserId:
                  type: string
                metadata:
                  type: object
      responses:
        '201':
          description: Vinculacion creada o actualizada
    get:
      summary: Resolver vinculacion externa a SVP
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: sourceApp
          required: true
          schema:
            type: string
        - in: query
          name: externalUserId
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Vinculacion encontrada
        '404':
          description: Vinculacion no encontrada
  /api/events/validate:
    post:
      summary: Validar payload de resultados externos sin aplicar puntos
      parameters:
        - in: header
          name: x-webhook-secret
          required: false
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [eventId, sourceApp, activityType, activityId, unit]
              properties:
                eventId:
                  type: string
                sourceApp:
                  type: string
                sourceEnv:
                  type: string
                svpUserId:
                  type: string
                externalUserId:
                  type: string
                activityType:
                  type: string
                activityId:
                  type: string
                score:
                  type: number
                  description: Puntaje de logro opcional (achievement score).
                activityHours:
                  type: number
                  description: Horas de actividad para formula SPV principal (hours * votes).
                totalVotes:
                  type: integer
                localVotes:
                  type: integer
                globalVotes:
                  type: integer
                unit:
                  type: string
                metadata:
                  type: object
      responses:
        '200':
          description: Payload validado y proyeccion de puntos
  /api/events/results:
    post:
      summary: Ingesta de resultados externos y acreditacion de puntos SVP
      parameters:
        - in: header
          name: x-webhook-secret
          required: false
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [eventId, sourceApp, activityType, activityId, unit]
              properties:
                eventId:
                  type: string
                sourceApp:
                  type: string
                sourceEnv:
                  type: string
                svpUserId:
                  type: string
                externalUserId:
                  type: string
                activityType:
                  type: string
                activityId:
                  type: string
                score:
                  type: number
                  description: Puntaje de logro opcional (achievement score).
                activityHours:
                  type: number
                  description: Horas de actividad para formula SPV principal (hours * votes).
                totalVotes:
                  type: integer
                localVotes:
                  type: integer
                globalVotes:
                  type: integer
                unit:
                  type: string
                requestId:
                  type: string
                occurredAt:
                  type: string
                  format: date-time
                metadata:
                  type: object
      responses:
        '201':
          description: Resultado procesado y puntos acreditados
        '200':
          description: Reintento idempotente
        '404':
          description: No se pudo resolver el usuario SVP
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT